+++ main.py
@@ -1,8 +1,10 @@
-from transport.node import Node, RingBuffer
+from transport.node import Node
+
+
 def message(_in_buffer, _out_buffer):
     print("Hey")
 
-# node = Node(id="12", in_size = 10, out_size= 10, action=message)
 
-# node.run_action()
+node = Node(id="12", in_size=10, out_size=10, action=message)
 
+node.run_action()
