+++ requirements.txt
@@ -1,3 +1,6 @@
 pytest
 pydantic
 mypy
+black
+pylint
+
