+++ transport/graph.py
@@ -1,4 +1,2 @@
 def add_nums():
     return 4
-
-
